refactor:  python files to module structure and ci conda-lock fix
