+++ .pre-commit-config.yaml
@@ -43,23 +43,18 @@ repos:
     rev: 2.12.4 # a PDM release exposing the hook
     hooks:
       - id: pdm-lock-check
+        args: [--lockfile, python/pdm.lock]
 
   - repo: local
     hooks:
       - id: sync-conda-lock
         name: sync-conda-lock
-        language: system
-        types: [python]
-        pass_filenames: false
-        always_run: true
+        language: python
+        verbose: true
+        files:  ^(\.pre-commit-hooks.yaml|environment.yml|conda-lock.yml|python/pyproject.toml)$
         entry: >
           bash -c 'cd python &&
-          python -c "import utils; utils.run_conda_lock(
-              replace_file_path=True,
-              default_kwargs=True,
-            )"'
-
-
+          python -c "from clim_recal.utils import _pre_commit_conda_lock; _pre_commit_conda_lock(execute_all=True)"'
 
   # Currently fails to build.
   # See: https://github.com/lorenzwalthert/precommit/issues/476
