+++ python/clim_recal/debiasing/debias_wrapper.py
@@ -6,7 +6,7 @@ from os import PathLike
 from pathlib import Path
 from typing import Final, Generator, Optional, Union
 
-from utils import (
+from ..utils import (
     DATE_FORMAT_SPLIT_STR,
     DATE_FORMAT_STR,
     DateType,
