+++ python/clim_recal/debiasing/preprocess_data.py
@@ -14,8 +14,8 @@ from pathlib import Path
 
 import numpy as np
 
-sys.path.insert(1, "../load_data")
-from data_loader import load_data
+#sys.path.insert(1, "../load_data")
+from ..data_loader import load_data
 
 # * ----- L O G G I N G -----
 formatter = logging.Formatter(
