+++ python/clim_recal/debiasing/three_cities_debiasing_cropped.sh
